requirements/stable.txt
djangorestframework==3.3.3
Django==1.9.4
